+++ src/server-middleware/controller/index.ts
@@ -1,0 +1,8 @@
+import { Router } from 'express'
+
+const router = Router()
+router.get('/hello', (_req, res) => {
+  res.json({ message: 'hello world' })
+})
+
+export default router
